+++ tools/infer/infer-pm-index256.py
@@ -112,7 +112,7 @@ def get_f0(x, p_len, f0_up_key=0):
     f0_mel[f0_mel <= 1] = 1
     f0_mel[f0_mel > 255] = 255
     # f0_mel[f0_mel > 188] = 188
-    f0_coarse = np.rint(f0_mel).astype(np.int)
+    f0_coarse = np.rint(f0_mel).astype(np.int32)
     return f0_coarse, f0bak
